+++ portal-impl/src/com/liferay/portlet/PortletBagFactory.java
@@ -67,6 +67,7 @@ import com.liferay.portal.kernel.xmlrpc.Method;
 import com.liferay.portal.notifications.UserNotificationHandlerImpl;
 import com.liferay.portal.util.JavaFieldsParser;
 import com.liferay.portal.util.PropsValues;
+import com.liferay.portlet.internal.FriendlyURLMapperTrackerImpl;
 import com.liferay.registry.Filter;
 import com.liferay.registry.Registry;
 import com.liferay.registry.RegistryUtil;
