+++ portal-impl/src/com/liferay/portlet/internal/FriendlyURLMapperTrackerImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet;
+package com.liferay.portlet.internal;
 
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringPool;
