+++ portal-impl/src/com/liferay/portlet/internal/RouteImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet;
+package com.liferay.portlet.internal;
 
 import com.liferay.portal.kernel.portlet.Route;
 import com.liferay.portal.kernel.util.InheritableMap;
