+++ portal-impl/test/integration/com/liferay/portlet/FriendlyURLMapperTrackerImplTest.java
@@ -22,6 +22,7 @@ import com.liferay.portal.model.impl.PortletImpl;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.SyntheticBundleRule;
 import com.liferay.portlet.bundle.friendlyurlmappertrackerimpl.TestFriendlyURLMapper;
+import com.liferay.portlet.internal.FriendlyURLMapperTrackerImpl;
 
 import org.junit.Assert;
 import org.junit.ClassRule;
