+++ portal-impl/test/unit/com/liferay/portlet/RouterImplTest.java
@@ -19,6 +19,7 @@ import com.liferay.portal.kernel.test.AssertUtils;
 import com.liferay.portal.kernel.util.HttpUtil;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.util.HttpImpl;
+import com.liferay.portlet.internal.RouterImpl;
 
 import java.util.HashMap;
 import java.util.Map;
